Standardize UI heights to 34px and resolve compilation issues
Implemented UI standardization across all panels as per Plan-59:
- Updated title bar and status bar heights to 34px in MainWindow, ContentPanel, CategoryPanel, NavPanel, MetaPanel, FilterPanel, and TagManagerView.
- Synchronized address bar component heights (32px) and mouse drag regions (34px).
- Fixed hardcoded QSS header heights in ContentPanel and BatchRenamePreviewDialog.
- Resolved compilation error by adding missing 'isManagedGlobal' member to RuntimeMeta struct in MetadataManager.h.
- Cleaned up compilation warnings: addressed size_t to int conversions in ContentPanel and unused parameter in DatabaseManager.
- Verified vertical centering of all header elements across the system.

+++ src/ui/ContentPanel.cpp
@@ -2120,7 +2120,7 @@ void ContentPanel::loadDirectory(const QString& path, bool recursive) {
         }
 
         std::vector<ItemRecord> allItems;
-        int lastEmittedCount = 0;
+        size_t lastEmittedCount = 0;
  
         std::function<void(const QString&, bool)> scanDir; 
         scanDir = [&](const QString& p, bool rec) { 
@@ -2139,7 +2139,7 @@ void ContentPanel::loadDirectory(const QString& path, bool recursive) {
                 if (recursive && (allItems.size() - lastEmittedCount >= 50)) {
                     lastEmittedCount = allItems.size();
                     QMetaObject::invokeMethod(panelPtr, "scanProgress", Qt::QueuedConnection, 
-                                             Q_ARG(int, allItems.size()), Q_ARG(QString, absPath));
+                                             Q_ARG(int, (int)allItems.size()), Q_ARG(QString, absPath));
                 }
  
                 if (rec && info.isDir()) { 
@@ -2174,7 +2174,7 @@ void ContentPanel::loadDirectory(const QString& path, bool recursive) {
                 if (recursive) {
                     qint64 elapsed = QDateTime::currentMSecsSinceEpoch() - startTime;
                     QMetaObject::invokeMethod(panelPtr, [panelPtr, allItems, elapsed]() {
-                        if (panelPtr) emit panelPtr->scanFinished(allItems.size(), elapsed);
+                        if (panelPtr) emit panelPtr->scanFinished((int)allItems.size(), elapsed);
                     }, Qt::QueuedConnection);
                 }
